+++ src/table/column/boolean/newEncode.js
@@ -1,10 +1,8 @@
 var purify = require('./purify');
 var newParam = require('../../query/newParameterized');
 var getSessionSingleton = require('../../getSessionSingleton');
-var newEncodeSafe = require('../newEncodeSafe');
 
 function _new(column) {
-
 
 	function encode(value) {
 		value = purify(value);
@@ -15,12 +13,10 @@ function _new(column) {
 		}
 		var encodeCore =  getSessionSingleton('encodeBoolean');
 
-		if (value)
-			return newParam(encodeCore(true));
-		return newParam(encodeCore(false));
+
+		return newParam('?', [encodeCore(value)]);
 	}
 
-	encode.safe = newEncodeSafe(column, purify);
 	return encode;
 }